+++ src/App.tsx
@@ -1,54 +1,17 @@
 import { css } from "@emotion/react";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import Picking from "./route/Picking";
-import SelectPreset, { TPreset } from "./route/SelectPreset";
+import SelectPreset from "./route/SelectPreset";
 import Upload from "./route/Upload";
 import Box from "./ui/Box";
 import Title from "./ui/Title";
 import Flexbox from "./ui/Flexbox";
-
-const LOCAL_KEY = "digipick-prev";
-const handleLocalStorageSaving = ({
-  uploadedPhoto,
-  preset,
-}: {
-  uploadedPhoto: string;
-  preset: TPreset;
-}) => {
-  const currentLocalStorage = localStorage.getItem(LOCAL_KEY);
-  const newItem = {
-    image: uploadedPhoto,
-    preset,
-  };
-  if (!currentLocalStorage) {
-    localStorage.setItem(LOCAL_KEY, JSON.stringify([newItem]));
-  } else {
-    const newStorage = JSON.stringify([
-      ...JSON.parse(currentLocalStorage),
-      newItem,
-    ]);
-    localStorage.setItem(LOCAL_KEY, newStorage);
-  }
-};
+import { useLocalStorage } from "./component/useLocalStorage";
 
 export default function App() {
   const [uploadedPhoto, setUploadedPhoto] = useState(null);
   const [preset, setPreset] = useState(null);
-  const [prevImages, setPrevImages] = useState([]);
-
-  useEffect(() => {
-    if (uploadedPhoto && preset) {
-      handleLocalStorageSaving({ uploadedPhoto, preset });
-    }
-  }, [uploadedPhoto, preset]);
-
-  useEffect(() => {
-    if (!uploadedPhoto && !preset) {
-      const currentLocalStorage = localStorage.getItem(LOCAL_KEY);
-      currentLocalStorage && setPrevImages(JSON.parse(currentLocalStorage));
-    }
-  }, []);
-
+  const prevImages = useLocalStorage({ preset, uploadedPhoto });
   return (
     <div
       css={css`
@@ -85,7 +48,7 @@ export default function App() {
                     margin-top: 20px;
                   `}
                 >
-                  {prevImages.slice(0, 3).map((prevImage) => (
+                  {prevImages.map((prevImage) => (
                     <button
                       css={css`
                         padding: 0;
